refactor: Migrate away from `grpc_utils` namespace
