+++ google/cloud/spanner/internal/session_pool.cc
@@ -16,8 +16,8 @@
 #include "google/cloud/spanner/internal/connection_impl.h"
 #include "google/cloud/spanner/internal/retry_loop.h"
 #include "google/cloud/spanner/internal/session.h"
-#include "google/cloud/grpc_utils/completion_queue.h"
-#include "google/cloud/internal/background_threads_impl.h"
+#include "google/cloud/background_threads.h"
+#include "google/cloud/completion_queue.h"
 #include "google/cloud/internal/make_unique.h"
 #include "google/cloud/log.h"
 #include "google/cloud/status.h"
